Updated vector and shape classes from float coordinates to double

+++ core/src/main/scala/Ball.scala
@@ -1,6 +1,6 @@
 package com.walter.eightball
 
 /** Represents a ball on the pool table */
-class Ball(var x: Float, var y: Float, val number: Int) extends Shape {
+class Ball(var x: Double, var y: Double, val number: Int) extends Shape {
   
 }
