+++ core/src/main/scala/Rectangle.scala
@@ -2,6 +2,6 @@ package com.walter.eightball
 
 /** Represents a rectangle formed shaped */
 trait Rectangle extends Shape {
-  var width: Float
-  var height: Float
+  var width: Double
+  var height: Double
 }
