+++ core/src/main/scala/Vector2D.scala
@@ -2,6 +2,6 @@ package com.walter.eightball
 
 /** Represents a vector in a 2D space */
 trait Vector2D {
-  var x: Float
-  var y: Float
+  var x: Double
+  var y: Double
 }
